review sysutils/code_80AE86C
