+++ include/ap.h
@@ -38,7 +38,7 @@ void AP_LoadDefinition(struct APHandle *, const u16 *);
 void AP_ExecDummyFrame(struct APHandle *);
 void AP_Init(struct APHandle *, const u16 *, u16);
 struct APHandle * AP_Find(const u16 * definition);
-ProcPtr APProc_Create(const void * apDefinition, int xPos, int yPos, int tileBase, int anim, u16 aObjNode);
+ProcPtr APProc_Create(const void * apDefinition, int xPos, int yPos, int tileBase, int anim, int aObjNode);
 // ??? APProc_OnUpdate(???);
 // ??? APProc_OnEnd(???);
 void APProc_SetParameters(struct APProc * proc, int x, int y, int tileBase);
